add abstract logger and implements channel logger for psr
