+++ src/LogFake.php
@@ -6,7 +6,7 @@ use Psr\Log\LoggerInterface;
 use Illuminate\Support\Collection;
 use PHPUnit\Framework\Assert as PHPUnit;
 
-class LogFake implements LoggerInterface
+class LogFake extends FakeLogger implements LoggerInterface
 {
     /**
      * All of the created logs.
@@ -157,102 +157,6 @@ class LogFake implements LoggerInterface
         return Collection::make($this->logs)->filter(function ($log) {
             return $this->currentChannelIs($log['channel']);
         });
-    }
-
-    /**
-     * Log an emergency message to the logs.
-     *
-     * @param  string  $message
-     * @param  array  $context
-     * @return void
-     */
-    public function emergency($message, array $context = [])
-    {
-        $this->log(__FUNCTION__, $message, $context);
-    }
-
-    /**
-     * Log an alert message to the logs.
-     *
-     * @param  string  $message
-     * @param  array  $context
-     * @return void
-     */
-    public function alert($message, array $context = [])
-    {
-        $this->log(__FUNCTION__, $message, $context);
-    }
-
-    /**
-     * Log a critical message to the logs.
-     *
-     * @param  string  $message
-     * @param  array  $context
-     * @return void
-     */
-    public function critical($message, array $context = [])
-    {
-        $this->log(__FUNCTION__, $message, $context);
-    }
-
-    /**
-     * Log an error message to the logs.
-     *
-     * @param  string  $message
-     * @param  array  $context
-     * @return void
-     */
-    public function error($message, array $context = [])
-    {
-        $this->log(__FUNCTION__, $message, $context);
-    }
-
-    /**
-     * Log a warning message to the logs.
-     *
-     * @param  string  $message
-     * @param  array  $context
-     * @return void
-     */
-    public function warning($message, array $context = [])
-    {
-        $this->log(__FUNCTION__, $message, $context);
-    }
-
-    /**
-     * Log a notice to the logs.
-     *
-     * @param  string  $message
-     * @param  array  $context
-     * @return void
-     */
-    public function notice($message, array $context = [])
-    {
-        $this->log(__FUNCTION__, $message, $context);
-    }
-
-    /**
-     * Log an informational message to the logs.
-     *
-     * @param  string  $message
-     * @param  array  $context
-     * @return void
-     */
-    public function info($message, array $context = [])
-    {
-        $this->log(__FUNCTION__, $message, $context);
-    }
-
-    /**
-     * Log a debug message to the logs.
-     *
-     * @param  string  $message
-     * @param  array  $context
-     * @return void
-     */
-    public function debug($message, array $context = [])
-    {
-        $this->log(__FUNCTION__, $message, $context);
     }
 
     /**
